# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewPeace BookStore API",
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "會員登入",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "信箱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登入成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "token": "aabbcc"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "登入失敗",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "invalid credentials"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "失敗(請求格式錯誤)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "bad request"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "會員註冊",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "信箱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sid",
                        "in": "query",
                        "description": "學號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "major",
                        "in": "query",
                        "description": "系別",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "註冊成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1,
                                    "token": "aabbcc"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "失敗(重複註冊)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "duplicate user"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "失敗(請求格式錯誤)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "bad request"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "取得登入資訊",
                "operationId": "98782e40ab294752958a37adbadd0a5d",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "id": 3,
                                    "name": "3",
                                    "email": "3",
                                    "role": 0,
                                    "sid": "3",
                                    "major": 1,
                                    "remember_token": null,
                                    "created_at": "2021-11-12T15:15:10.000000Z",
                                    "updated_at": "2021-11-12T15:15:10.000000Z"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "失敗(Token錯誤)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "登出",
                "operationId": "a8dd056be3e98a38c657f73ed6cdef2d",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/bindPhone": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "綁定手機",
                "operationId": "cb54454c80430135089554abb6e254bb",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "手機號碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "失敗(已綁定手機)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "already bind phone"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "失敗(手機號碼錯誤)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "phone format error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/verifyPhone": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "驗證手機",
                "operationId": "2ecc089d9963d342bb6b4190ce47fcdc",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "驗證碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 1
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "失敗(未綁定手機)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "not bind phone yet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "失敗(已驗證成功)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "had verify phone"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "失敗(驗證碼錯誤)",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": 0,
                                    "message": "error code"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}